face_embedding_service/requirements.txt
numpy==1.24.3
pandas==2.2.2
Pillow==10.4.0
uvicorn==0.15.0
fastapi==0.96.0
opencv-python==4.10.0.84
tensorflow[and-cuda]==2.16.2
requests==2.32.3
gdown==5.2.0
tf-keras==2.16.0
python-multipart==0.0.9